COP2800GroupWork (completed)
I don't think we can link this all together, so I recommend we send it to the prof. in a compressed .zip with all the classes of each assignment. I will exchange the folder to all of you so we can turn it in. If I don't see anything that we can link this program, then I will be sending it through .zip. 

-Nick

P.S The assignment is due by 4:24 PM today, so I suggest we hurry